molmod/requirements.txt
certifi==2023.7.22
chardet==5.0.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.3.2
Flask-WTF==1.2.0
Flask-Mail==0.9.1
idna==3.4
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.3
MarkupSafe==2.1.1
requests==2.31.0
urllib3==1.26.18
Werkzeug==3.0.1
WTForms==3.0.1
pandas==1.5.0
psycopg2-binary==2.9.3
openpyxl==3.0.10
Flask-CAS==1.0.2
csvkit==1.1.1
sqlalchemy<2.0
bs4==0.0.2
